+++ tests/test_contextmanager.py
@@ -1,15 +1,16 @@
 import asyncio
 
 import pytest
+from async_generator import yield_
 
-from asyncio_extras import yield_async, async_contextmanager
+from asyncio_extras import async_contextmanager
 
 
 @pytest.mark.asyncio
 async def test_async_contextmanager():
     @async_contextmanager
     async def dummycontext(value):
-        await yield_async(value)
+        await yield_(value)
 
     async with dummycontext(2) as value:
         assert value == 2
@@ -20,11 +21,29 @@ async def test_async_contextmanager_three_awaits():
     @async_contextmanager
     async def dummycontext(value):
         await asyncio.sleep(0.1)
-        await yield_async(value)
+        await yield_(value)
         await asyncio.sleep(0.1)
 
     async with dummycontext(2) as value:
         assert value == 2
+
+
+@pytest.mark.asyncio
+async def test_async_contextmanager_exception():
+    @async_contextmanager
+    async def dummycontext(value):
+        nonlocal exception
+        try:
+            await yield_(value)
+        except Exception as e:
+            exception = e
+
+    exception = None
+    with pytest.raises(Exception):
+        async with dummycontext(2):
+            raise Exception('foo')
+
+    assert str(exception) == 'foo'
 
 
 @pytest.mark.asyncio
@@ -33,19 +52,17 @@ async def test_async_contextmanager_no_yield():
     async def dummycontext():
         pass
 
-    with pytest.raises(RuntimeError) as exc:
+    with pytest.raises(StopAsyncIteration):
         async with dummycontext():
             pass
-
-    assert str(exc.value) == 'coroutine finished without yielding a value'
 
 
 @pytest.mark.asyncio
 async def test_async_contextmanager_extra_yield():
     @async_contextmanager
     async def dummycontext(value):
-        await yield_async(value)
-        await yield_async(3)
+        await yield_(value)
+        await yield_(3)
 
     with pytest.raises(RuntimeError) as exc:
         async with dummycontext(2) as value: